[animation] create animation-fundamentals module to share common code.
