+++ lib/js/components/layouts/example-key-moments.mjs
@@ -8,7 +8,7 @@ import {
   // , unwrapPotentialWriteProxy
   , StateComparison
   , CoherenceFunction
-  , BooleanModel
+  // , BooleanModel
   , StringModel
   , NumberModel
   , ValueLink
@@ -38,8 +38,7 @@ import {
 } from '../basics.mjs';
 
 import {
-    UIToggleButton
-  , StaticTag
+    StaticTag
   , StaticNode
   , DynamicTag
   , UINumberInput
@@ -50,20 +49,15 @@ import {
 } from '../generic.mjs';
 
 import {
-    AnimationTGenerator
-} from '../animation-t-generator.mjs';
-
-import {
     UIManualAxesLocations
 } from '../ui-manual-axis-locations.mjs';
 
 import {
-    UITimeControlKeyMomentSelectCircle
-} from '../ui-time-control-circle.mjs';
-
-
-
-
+    timeControlMixin
+  , AnimationTGenerator
+  , UITimeControlKeyMomentSelectCircle
+  , basicPlayerControlWidgets
+}  from '../animation-fundamentals.mjs';
 
 const FontSizeModel = _AbstractSimpleOrEmptyModel.createClass(NumberModel)
   , KeyMomentModel = _AbstractStructModel.createClass(
@@ -134,35 +128,7 @@ const FontSizeModel = _AbstractSimpleOrEmptyModel.createClass(NumberModel)
   , ActiveKeyMomentsModel = _AbstractListModel.createClass('ActiveKeyMomentsModel', ActiveKeyMomentModel)
   , ExampleKeyMomentsLayoutModel = _BaseLayoutModel.createClass(
         'ExampleKeyMomentsLayoutModel'
-        , CoherenceFunction.create(
-                  [/*'keyMoments'*/'t', 'playing', 'duration', 'isLoop', 'perpetual'],
-                function prepare(
-                  {/*keyMoments*/t, playing, duration, isLoop, perpetual}) {
-            // if(keyMoments.size === 0) {
-            //     // we could also push two initial keyframes as it is the
-            //     // basis for an animation, one keyframe is just static.
-            //     // TODO: the _AbstractListModel should have method to
-            //     //       create a primalState of its members directly.
-            //     keyMoments.push(keyMoments.constructor.Model.createPrimalState(keyMoments.dependencies));
-            // }
-            if(t.value === undefined)
-                t.value = 0;
-            if(playing.value === undefined)
-                playing.value = false;
-            if(isLoop.value === undefined)
-                isLoop.value = true;
-            if(perpetual.value === undefined)
-                perpetual.value = true;
-            if(duration.value === undefined)
-                duration.value = 1;// 1 second
-        })
-      , CoherenceFunction.create(
-            ['playing', 'duration'], function protectPlaying({playing, duration}) {
-            // NOTE: This is actually a nice use case for coherence functions.
-            if(duration.value === 0) {
-                playing.value = false;
-            }
-        })
+      , ...timeControlMixin
       , CoherenceFunction.create(
             ['activeKeyMoments'], function sortActiveKeyMoments({activeKeyMoments}) {
             // NOTE: it's not terrible important for most operations to have this
@@ -196,12 +162,6 @@ const FontSizeModel = _AbstractSimpleOrEmptyModel.createClass(NumberModel)
         })
       , ['keyMoments', KeyMomentsModel]
       , ['activeKeyMoments', ActiveKeyMomentsModel]
-        // need more controls
-      , ['t', NumberModel]
-      , ['duration', NumberModel] // in seconds
-      , ['isLoop', BooleanModel] // connect end with beginning and transition
-      , ['perpetual', BooleanModel] // never stop playback
-      , ['playing', BooleanModel]
     )
   ;
 
@@ -1665,53 +1625,7 @@ class ExampleKeyMomentsController extends _BaseContainerComponent {
               , activeKeyMomentsSortedKeys
               , _setActiveKeyMomentsValues
             ]
-          , [
-                {zone: 'main'}
-              , [
-                    ['playing', 'boolean']
-                ]
-              , UIToggleButton
-              , 'playing'
-              , 'pause'
-              , 'play'
-              , 'Toggle animation play/pause.'
-            ]
-          , [
-                // This should not be to control to play repeatedly
-                // or not. Loop !== play repeatedly.
-                {zone: 'main'}
-              , [
-                    ['isLoop', 'boolean']
-                ]
-              , UIToggleButton
-              , 'is-loop'
-              , 'is a loop'
-              , 'is end to end'
-              , 'Connect end with start and transition.'
-            ]
-          , [
-                // This should not be to control to play repeatedly
-                // or not. Loop !== play repeatedly.
-                {zone: 'main'}
-              , [
-                    ['perpetual', 'boolean']
-                ]
-              , UIToggleButton
-              , 'perpetual'
-              , 'play once'
-              , 'repeat'
-              , 'Toggle perpetual playback.'
-            ]
-          , [
-                {zone: 'main'}
-              , [
-                    ['duration', 'value']
-                ]
-              , UINumberInput
-              , 'Time Duration' // label
-              , 'seconds'// unit
-              , {min:0} // minMaxValueStep => set attribute
-            ]
+          , ... Object.values(basicPlayerControlWidgets)
           , [
                 {zone: 'main'}
               , [
